+++ module/applications/actor/base-sheet.mjs
@@ -628,7 +628,7 @@ export default class ActorSheet5e extends ActorSheetMixin(ActorSheet) {
       if ("weight" in item.system) ctx.totalWeight = ((item.system.quantity ?? 1) * item.system.weight).toNearest(0.1);
 
       // Weapon Firing Arc
-      if ("firingArc" in item.system) ctx.firingArc = CONFIG.SW5E.weaponFiringArcs[item.system.firingArc];
+      if ("firingArc" in item.system) ctx.firingArc = CONFIG.SW5E.weaponFiringArcs[item.system.firingArc]?.label;
 
       // Item properties
       ctx.propertiesList = item.propertiesList;